map_to_path: Sort cones before creating splines.
If the cones are unordered at the time splines are created, the splines will be incorrect resulting in an incorrect path. Therefore the cones must be sorted, starting from the far orange cone and ending at the close orange cone. This commit as an implementation of the sorting algorithm two-opt and uses it to sort the cones before the splines are created.