Python: Add common QoS profiles and use them in all subscribers
QoS policies give us control over how publishers and subscribers manage messages up until they are processed. Internally, subscribers store the messages they receive in a queue (the data structure), when the subscriber finishes processing a message, it processes the next message in the queue. As queues are FIFO, the subscriber processes older messages before newer messages, this is not always the behaviour we want. For example, when using the car's pose to calculate the next steering angle, we only ever want to be using the latest pose. The default behaviour of a subscriber is to store the previous 10 messages, in this example, that would mean we calculate the steering angle for where the car was 10 messages ago, and apply it to where the car currently is, leading to the car following a delayed version of the track.